web: blueprint
